Refactor complementary color calc
