+++ src/util/index.ts
@@ -1,3 +1,4 @@
 export { getLegibleTextColor } from "./getLegibleTextColor";
 export { hexToRgb } from "./hexToRgb";
 export {rgbToHex } from "./rgbToHex.ts";
+export {handleComplementaryColors} from "./handleComplementaryColors.ts"
